# public/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Lumina Slides API",
        "description": "API to generate and save Lumina Slides presentations, and retrieve schema definitions.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://us-central1-lumina-slides.cloudfunctions.net",
            "description": "Production Server"
        }
    ],
    "paths": {
        "/saveDeck": {
            "post": {
                "operationId": "saveDeck",
                "summary": "Save a new presentation deck",
                "description": "Accepts a JSON definition of a slide deck and saves it to Firestore, returning a unique ID and shareable URL.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Deck"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Deck saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "description": "The unique ID of the saved deck."
                                        },
                                        "url": {
                                            "type": "string",
                                            "description": "The direct URL to view the saved deck."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/updateNotes": {
            "post": {
                "operationId": "updateNotes",
                "summary": "Update speaker notes for specific slides",
                "description": "Updates only the speaker notes for the specified slide indices in an existing deck.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "deckId",
                                    "notes"
                                ],
                                "properties": {
                                    "deckId": {
                                        "type": "string",
                                        "description": "The ID of the deck to update"
                                    },
                                    "notes": {
                                        "type": "array",
                                        "minItems": 1,
                                        "items": {
                                            "type": "object",
                                            "required": [
                                                "slideIndex",
                                                "note"
                                            ],
                                            "properties": {
                                                "slideIndex": {
                                                    "type": "integer",
                                                    "minimum": 0,
                                                    "description": "The zero-based index of the slide to update"
                                                },
                                                "note": {
                                                    "type": "string",
                                                    "description": "The new markdown content for the speaker notes"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Notes updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "updatedSlides": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation Failed"
                    },
                    "404": {
                        "description": "Deck not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/patchDeck": {
            "post": {
                "operationId": "patchDeck",
                "summary": "Patch an existing deck (Partial Update)",
                "description": "Updates metadata and/or replaces specific slides without re-submitting the entire deck.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "deckId"
                                ],
                                "properties": {
                                    "deckId": {
                                        "type": "string"
                                    },
                                    "meta": {
                                        "type": "object",
                                        "properties": {
                                            "title": {
                                                "type": "string"
                                            },
                                            "author": {
                                                "type": "string"
                                            },
                                            "theme": {
                                                "type": "string",
                                                "description": "Preset theme name (e.g. 'cyber', 'ocean', 'luxury')"
                                            },
                                            "themeConfig": {
                                                "$ref": "#/components/schemas/ThemeConfig"
                                            },
                                            "orbColor": {
                                                "type": "string",
                                                "description": "Hex color for the global ambient orb (e.g. '#3b82f6')"
                                            }
                                        }
                                    },
                                    "slides": {
                                        "type": "array",
                                        "items": {
                                            "type": "object",
                                            "required": [
                                                "index",
                                                "content"
                                            ],
                                            "properties": {
                                                "index": {
                                                    "type": "integer",
                                                    "minimum": 0
                                                },
                                                "content": {
                                                    "description": "The new slide object (Statement, Half, Flex, etc.)",
                                                    "type": "object",
                                                    "additionalProperties": true
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Deck patched successfully"
                    },
                    "400": {
                        "description": "Validation Failed"
                    },
                    "404": {
                        "description": "Deck not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/getDeck": {
            "get": {
                "operationId": "getDeck",
                "summary": "Get an existing presentation deck",
                "description": "Retrieves the full JSON structure of a deck using its unique ID. Essential for identifying slide indices before patching.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "description": "The unique ID of the deck",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Full deck object",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Deck"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Deck not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Deck": {
                "type": "object",
                "properties": {
                    "meta": {
                        "type": "object",
                        "description": "Metadata about the presentation",
                        "properties": {
                            "title": {
                                "type": "string"
                            },
                            "author": {
                                "type": "string"
                            },
                            "theme": {
                                "type": "string",
                                "description": "Preset theme name"
                            },
                            "themeConfig": {
                                "$ref": "#/components/schemas/ThemeConfig"
                            },
                            "orbColor": {
                                "type": "string",
                                "description": "Hex color for the ambient orb"
                            }
                        }
                    },
                    "slides": {
                        "type": "array",
                        "description": "List of slides in the deck",
                        "items": {
                            "type": "object",
                            "description": "A generic slide object",
                            "properties": {
                                "type": {
                                    "type": "string",
                                    "description": "The type of slide layout"
                                }
                            }
                        }
                    }
                }
            },
            "ThemeConfig": {
                "type": "object",
                "properties": {
                    "colors": {
                        "type": "object",
                        "properties": {
                            "primary": {
                                "type": "string"
                            },
                            "background": {
                                "type": "string"
                            },
                            "text": {
                                "type": "string"
                            },
                            "muted": {
                                "type": "string"
                            }
                        }
                    },
                    "fonts": {
                        "type": "object",
                        "properties": {
                            "heading": {
                                "type": "string"
                            },
                            "body": {
                                "type": "string"
                            },
                            "mono": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    }
}